+++ src/commands/alias.cpp
@@ -2,9 +2,7 @@
 
 #include <string>
 
-int cmds::alias(const std::vector<std::string>& arg,
-                StreamManager& stream,
-                VariableManager& vars) {
+int cmds::alias(const std::vector<std::string>& arg, VariableManager& vars) {
     // add command
     size_t found = arg.at(1).find("=");
     std::string cmd_name = arg.at(1).substr(0, found);