+++ src/commands/clear.cpp
@@ -2,9 +2,7 @@
 
 #include <string_view>
 
-int cmds::clear(const std::vector<std::string>& arg,
-                StreamManager& stream,
-                VariableManager& vars) {
+int cmds::clear(const std::vector<std::string>& arg, VariableManager& vars) {
     std::string_view sys = vars.get("SYSTEM");
     if (sys != "Linux" && sys != "MacOS" && sys != "Unix") {
         system("CLS");