+++ src/shell.cpp
@@ -1,9 +1,13 @@
 #include "shell.h"
 
+#include <iostream>
 #include <string_view>
 #include <unistd.h>
 #include <stdlib.h>
 #include <signal.h>
+
+#include <fmt/format.h>
+
 #include "feature/path_str_gen.h"
 #include "feature/string_parser.h"
 #include "feature/theme.h"
@@ -11,8 +15,7 @@
 
 extern char** environ;
 
-Shell::Shell(std::istream& in, std::ostream& out, std::ostream& err)
-    : stream(in, out, err) {
+Shell::Shell() {
     vars.set("SYSTEM", sys)
         .set("SHELL", "yush");
 
@@ -44,10 +47,10 @@ int Shell::run(bool output) {
         }
 
         if (signal(SIGINT, signal_handler) == SIG_ERR) {
-            stream.err() << "Error: signal handler failed\n";
+            // stream.err() << "Error: signal handler failed\n";
         }
 
-        std::getline(stream.in(), input);
+        getline(std::cin, input);
 
         input = preprocess_cmd(input);
         std::vector<std::string> arg = string_parser(input, ' ');
@@ -60,18 +63,15 @@ int Shell::run(bool output) {
         }
 
         runtime_status = exec_cmd(input, arg);
-    } while (!stream.in().eof());
+    } while (!std::cin.eof());
 
     return runtime_status;
 }
 
 int Shell::output() {
-    stream.out() << "\n"
-                 << vars.get("USER") << "@"
-                 << vars.get("NAME") << ' '
-                 << path_str_gen(vars.get("HOME")) << '\n';
+    fmt::print("\n{}@{}{}\n", vars.get("USER"), vars.get("NAME"), path_str_gen(vars.get("HOME")));
     // stream.out() << runtime_status; // when debug
-    stream.out() << "> ";
+    fmt::print("> ");
     return 0;
 }
 
@@ -101,20 +101,15 @@ std::string Shell::preprocess_cmd(const std::string& cmd) {
 int Shell::exec_shell_builtin(const std::vector<std::string>& arg) {
     using namespace cmds;
     static const std::unordered_map<std::string, decltype(&alias)> command_map{
-        {"alias", alias}, {"cat", cat},
-        {"cd", cd},       {"clear", clear},
-        {"cp", cp},       {"date", date},
-        {"echo", echo},   {"function", function},
-        {"la", la},       {"ls", ls},
-        {"mkdir", mkdir}, {"mv", mv},
-        {"pwd", pwd},     {"rm", rm},
-        {"set", set},     {"touch", touch},
-        {"whoami", whoami},
+        {"alias", alias},       {"cd", cd},
+        {"clear", clear},       {"echo", echo},
+        {"function", function}, {"ls", ls},
+        {"pwd", pwd},           {"set", set},
     };
 
     auto command_it = command_map.find(arg[0]);
     if (command_it != command_map.cend()) {
-        return command_it->second(arg, stream, vars);
+        return command_it->second(arg, vars);
     }
 
     return 127;
@@ -152,7 +147,7 @@ int Shell::exec_cmd(const std::string current_command,
         }
     }
 
-    stream.err() << "command `" << arg[0] << "` not found.\n";
+    // stream.err() << "command `" << arg[0] << "` not found.\n";
 
     return 127;
 }